server/requirements.txt
asgiref==3.8.1
black==24.10.0
boto3==1.35.43
botocore==1.35.43
click==8.1.7
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
